[terminal] Specify terminal product configuration on top of base.gni
This specifies the 'terminal' product directly on top of the base
product definition (skipping JSON). It also moves the packages which
are needed at startup but aren't in /system or required to be the
monolith into the 'preinstall' package set. This allows iterating on
these components without needing to reboot.

I've left packages which contribute files to the
deprecated_system_image in the 'monolith' set, even though the build is
smart enough to pull these parts out, to try to make it clearer to
developers why changes to those sets require a full system update to
change.

Test: 'fx set ... --product terminal'

Change-Id: I46eb4f4a580157ee7fa9bcc05dd9cc4d0711acb2