[fbl][containers] Add support for std::unique_ptr
std::unique_ptr is coming to Zircon!  Add support to the fbl intrusive
containers for the std:: version of unique pointer.  Make sure to support
pointers which use custom deleters.

Also, extend the test framework to test both the custom deleter and non-custom
deleter forms of std::unique_ptr.  Finally, add some checks in the test
framework to make sure the custom deleter is being called the number of times we
would expect given the behavior of each test.

Test: Build and newly extended unit tests.
Change-Id: I60e3abb8315391fd85949eca9ca5daf9fb58fd6f