[fidlc] Deferring Type Construction Post Raw to Flat Conversion
Introducing a generic representation named TypeConstructor to capture
all areas where we reference types. This makes it possible to have a
single representation, share code around it, and most importantly, defer
type construction to a later stage in the compiler, once all constants
have been resolved.

For instance, with the example:

    const uint32 MAX_SIZE = 1024;
    struct MyStruct {
      string:MAX_SIZE desc;
      string:1024 other_desc;
    };

We must first resolve the MAX_SIZE constant to understand that the type
for both desc field, and other_desc field are the same.

Type constructors create types using type templates, which are looked up
by name. So for instance, the name "uint32" looks up a type template,
which is an instance of PrimitiveTypeTemplate, and which knows how to
instantiate a PrimitiveType representing a uint32.

All this machinery moves us closer to 1. canonicalizing types (because
their creation is now centralized, and soon easy to memoize), and 2.
introducing user-defined templates such as

    struct pair<L, R> { ....

Which would get materialized dynamically.

This also makes it natural to extend type aliasing to use any types, not
just primitives.

Test: make tools -j12 HOST_USE_ASAN=true && ./build-x64/host_tests/fidl-compiler-test
Change-Id: Id97c6751b8944b5ee04ff813d2cd0dcc47711f17