[fidl][rust][transitional] Test method soft transitions
This tests three versions of interface implementations against three
versions of a FIDL library. It does that by declaring the
implementations in macros that take the name of the FIDL library module
as an argument.

The code isn't meant to actually be run - it's just to test that it
compiles correctly.

TEST=built with //garnet/packages/tests/all

Change-Id: I55bd40469c12fba8faec8a4a404fa6602ccf49ad