[wlan][broadcom] Fix scanning protected APs
Broadcom driver doesn't keep the capability info. Because the privacy
bit is not set, our SME considers APs in the scan result as not
supported.

Strictly speaking, only preserving the capability info in the rx path is
necessary to fix the bug, but I have also made the change to preserve it
in the tx path.

WLAN-853 #done

Test: manually scan and connect with Broadcom device
Change-Id: I65f9b83c072d36a45840ebdeaeab2e730e503109