[musl][crashsvc] Noisily fail when IO functions are not implemented in libc.
Currently, if an app does not link in an implementation for POSIX IO
functions e.g. read/write, they automatically fallback to a dummy
implementation, which does nothing except silently fail.

This change explores if it is better to just crash in those scenarios.
The rationale is that in production, we should not be depending on these
dummy implementations. And currently because of the build system, it is
quite easy to forget to explicitly link in an implementation e.g. FDIO.

After opting to explicitly crash, the following cases have been
uncovered:
1. system/utest/util/listnode does not link FDIO. Hence it has no
console output whatsoever.
2. system/core/crashsvc also does not link FDIO, but calls fprintf.
Hence when we crash on unimplemented functions, it will forever
recursively crash in the crash handler. This has also been fixed.

TEST: Run all tests with the crashing behavior. And also try randomly
removing FDIO from any test and verifies that it noisily crashes.

Global Integration Test: integration/+/78875

Change-Id: I710e512068c2926f8ff09a2b71c806b7bf6f87df