[wlan][rsn] Relax IV constraint in Group Handshake
IEEE Std 802.11-2016, 12.7.7.2 requires EAPOL key frame's IV to be
zeroed when EAPOL protocol version is following 802.1X-2004, and
random if 802.1X-2001 is used.
Some routers in our tests violate this requirement and as a result
Group Keys cannot be re-keyed. This change relaxes this
requirement. That's safe as the IV is ignored in the handshake.

WLAN-849 #done

Test=Added tests. Verified compatibility with a couple of 3rd
party APs.

Change-Id: I0e4f44d298ab8fc2d02ff1a6eb6b718c9205c19c