[session_shell] Time out on slow SessionShell
DetachView() must not block the Stop() operation indefinitely.
The StopOperation must time out when DetachView() takes too long.

Added a test case to test the behavior. Unfortunate about tests
for timeout behavior is they need the time until out to elapse,
i.e. such a test is slow. In this case, the timeout in question
is 1 second, which is bearable.

Also changed the behavior such that on logout, no DetachView()
calls are made for stories that are still running, in accordance
with the behavior documented in session_shell.fidl. Added a
test case for that too.

TESTED=Added test cases to session_shell integration test.

MF-121 #comment Time out on slow SessionShell.

Change-Id: Iaaec29df7d6edb2fb8ce5269ff1c02d423f19555