[bt][gap] Accept BR/EDR link key in RemoteDeviceCache
Add a link key member to RemoteDevice::BrEdrData.

Introduce a StoreBrEdrBond that stores a link key to a record of a
known device and makes it bonded.

Update logic for RemoteDevice::bonded().

Extract cache insertion logic into an InsertDevice function that is also
refactored to use std::unordered_map::try_emplace and structured
bindings.

BT-424
Test: Add GAP_RemoteDeviceCacheTest_BondingTest cases to bt-host-unittests

Change-Id: Ife99a91d295d598f9058ec8519a201fe60c86a09