[wayland][bridge] Use internal mutability for event counter.
In an upcoming change I'll be iterating over a set of object IDs in the
Client and emitting input events for each. In that context I'll need to
either clone the set of id's I'm interating over, or remove the
requirement to have a mutable Client reference to increment the counter.

I initially factored this into a clone-able `EventCounter` struct that
wrapped an Rc<Cell<u32>>, but that overhead seems unnecessary so I just
inlined the Cell directly into the Client.

Test: CQ
Change-Id: I75308ab215bde3d3392285d503844cd47fb9c17c