[netstack] Simplify dhcp context management
...by moving the cancel function into the embedded dhcp state struct and
clarifying the context's semantics in the process. The context is now
initialized when a NIC transitions to eth.StateStarted and cancelled
when either DHCP is disabled or the NIC transitions to eth.StateStopped.

Also add unit tests of DHCP/static IP state management across restarts.

Bug: NET-1855 #comment unit tests for DHCP on device.
Tested: fx run-image-test netstack_test
- fx run-test netstack_integration_test
- with edgerouter bridged to qemu,
  - fx run -kN
  - ifconfig ethp0003 dhcp stop
  - ifconfig ethp0003 dhcp start
    - output:
      DHCP acquired IP 192.168.4.91 for 24h0m0s
      Adding DNS servers: [8.8.8.8]

Change-Id: Ic3b24b2f342a59c531f9e44093fbeea54ce01152