[linux_guest] Fix init script location and mounting
Previously linux was not finding our init script at all and defaulting
to just running /bin/sh. This places it somewhere that it should be
found (/etc/init). Also we should mount /sys, /proc, and the test utils.

Also clean up the script to remove CPIO images and related flags.

TEST=Boot linux_guest and check for mounted directories

Change-Id: I4cb3f24e3128f3e3ea8ea6e48357c40ffd6d4add