[bt][fbl] Remove std::unique_ptr traits from BT stack.
BT was using fbl:: containers, but wanted to use std::unqiue_ptr
instead of fbl::unique_ptr.  Because of this, they needed to carry
around their own definition of the pointer traits in order for the
containers to be usable.

Zircon now has support for std::unique_ptr now, however.  So the
containers support it by default, and we no longer need extra code up
at this level in order to allow this usage pattern.

Test: Build and runtests on a VIM2
Change-Id: I7204a0cab5b6728d7f6c202f1bf21398dc91c7e3