[debugger] arm64 debug agent integration.
This is the low level helper tests that modify the
zx_thread_state_debug_regs_t values to the correct state whether we're
adding or removing HW breakpoints.

Better logging and test cases for x64 too.
HW breakpoint setting for x64 is covered by the integration breakpoint
tests. arm64 tests for that are not activated yet as more hooking up is
still needed.

TEST=unit

Change-Id: I1af8b49c6555a10763a54e96bd88086c1098465d