[gce] Use pigz if available to compress fx disk image
fx gce create-fuchsia-image constructs a fuchsia disk image suitable
for GCE; the image is a raw disk image inside a gzipped tarball.
One of the slower steps of constructing the disk image is gzipping
the tarball; use pigz (parallel gzip) if available to save time.

On a Xeon 'Skylake' 24 vCPU GCE instance, reduces the time for
compression in create-fuchsia-image from 2 min to 9 seconds.

Change-Id: Ic880890737f6af585f6e91406655097cd79dbab6