[zxtest]: Initial commit.
Introduce Test class, which is the base for all fixtures.

Introduce TestDriver as an interface for propagating failures
and tracking the state of a test.

Added tests under zxtest/test. To avoid going full meta, and for sanity,
these tests rely on runtiome assertions to verify that the promised
behavior is correct, and are manually registered. This test, has no
output yet, it will crash if anuy assertion fails.

TEST=zxtest-test

Change-Id: I79eaa1377ea76f8fcf6938c95107aea62c9cb27a