[audio] Rename fuchsia.media.Audio to AudioCore
As we add more functionality to audio_core, it will be migrating
into becoming a priviledged service that clients will access
through higher layer APIs indirectly.  To maintain compatability
as we transition through this process, rename the interface to
AudioCore, and create a separate shim service that implements
the old API set.

Once the final Client APIs have landed and all users of the shim
have migrated to it, the shim service can go away.

TEST: fx run-test audio_fidl_tests
      fx run-test audio_mixer_tests
      fx run-test mediaplayer_tests
Change-Id: Ifbe40457916aee2b64187f6ddb4ed5043f79960a