[bt][gap] Identify dual-mode remote devices
Make RemoteDeviceCache::FindDeviceByAddress search by public identity
address of both BR/EDR and LE technologies.

Disallow creation of a new device or new bond if it would result in
duplicate entries for BR/EDR and LE with the same address values.

When upgrading a device to dual-mode, add the new technology's identity
address to the address cache.

When removing a device, release all of its known addresses, including
those for LE devices employing address privacy.

Add an |address_| field and getter to RemoteDevice::BrEdrData for the
BD_ADDR as a DeviceAddress of BR/EDR type.

Minor code cleanups.

BT-629
BT-630
Test:
 - GAP_RemoteDeviceCacheTest.*DeviceBecomesDualMode*
 - GAP_RemoteDeviceCacheTest.LookUp*DeviceBy*Alias
 - GAP_RemoteDeviceCacheTest.DisallowNew*DeviceIf*DeviceExists
 - GAP_RemoteDeviceCacheTest.AddBondedLowEnergyDeviceFailsWithExistingBrEdrAliasAddress
 - GAP_BrEdr*ManagerTest.*UpgradesKnownLowEnergyDeviceToDualMode
 - GAP_LowEnergy*ManagerTest.*UpgradesKnownBrEdrDeviceToDualMode

Change-Id: Ife02669bc4c4054bffd5719dc2862563dde5dce3