[kernel][vm] Implement commit on pager vmos
Committing a pager vmo generates page requests to the backing pager
service. The kernel will attempt to batch requests for adjacent pages,
but the pager API makes no guarantees about exactly what the batching
behavior is.

For a single commit operation, the current implementation generates a
single batched page request for each run of non-resident pages which do
not already have outstanding page requests. For a commit operation
which is broken up into multiple requests, the requests are dispatched
serially, in the sense that the next request will not be dispatched
until the current request is fully supplied.

This implementation uses a missing page count to track how much work is
left for a pending page request. This can result in redundant page
requests if pages are removed from under a pending page request. Given
that it greatly simplifies request tracking and that redundant requests
are not technically a pager API violation, it seems like a worthwhile
trade-off. Furthermore, this should be rare in the current system, and
may become impossible if decommit becomes less forceful (and evict
handles it).

The commit-centric pager tests in core-tests do test the batching
behavior of the kernel, as the correctness of such behavior is
important for the efficiency of the system. However, that does not mean
that the behavior is part of the pager API's contract.

Test: /boot/test/core/pager-test
Test: Pave, boot, use session shell on nuc
Test: vmstress from sandbox/stevensd/userpager for 48+ hours on
nuc/vim
Test: repeatedly run and kill vmstress for 12+ hours on nuc
ZX-1480 #comment [kernel][vm] Implement commit on pager vmos

Change-Id: I65b594a9f3957f2d1e94086852fb3e322847afcd