[syscalls][docs] Cross-link syscall documentation
This uses the format [`zx_syscall()`] inline, and defines a reference at
the bottom of the file, to avoid having to stutter the syscall name both
as inline text and as the .md file in () after the []. The references at
the bottom of the file are not included in rendered output. (Thanks
bruce.mitchener for the suggested format.)

Additionally, replace existing spellings of syscalls that vary in style,
sometimes prefixing with zx_, sometimes including () inside bold,
sometimes both, and sometimes neither. On a syscall's own page, it is
not linked, and is instead simply `zx_syscall()`. There are some
existing references to syscalls (primarily those that were already
linked) that are not updated by this. Maybe later, or manually.

See the gitiles link for channel_call.md for an example result.

ZX-3106 #comment [syscalls][docs] Cross-link syscall documentation

Test: CQ
Change-Id: Icac7c50251012af3580c130553d1bfd640fb3eff