[debugger] De-virtualize Process- and TargetSymbols
The ProcessSymbols and TargetSymbols classes are holders and aggregators
for a set of modules. This makes them not very useful to mock. There
were no other implementations of TargetSymbols, and the several uses of
MockProcessSymbols weren't doing very much.

The users of MockProcessSymbols were moved to the
ProcessSymbolsTestSetup class (renamed without "Impl" now) so there are
fewer ways to setup tests using mock symbol information.

Some additions were made to MockModuleSymbols that MockProcessSymbols
previously implemented.

TEST=unit

Change-Id: Ifd2d6644a897eb51339fb796ee1705baf95a742b