[dump_breakpad_symbols] Have dump_breakpad_symbols produce an archive
The go tool itself should do the tarring, as it easiest to do there where
the tool has the relevant information; this would save the recipe from
having to make brittle assumptions. If the build produced the archive directly,
the recipe would have to do no more work beyond just uploading it, to the same downstream effect.

IN-851 #comment
Test: ninja -C out/x64 build/gn:breakpad_symbols #The desired tarball is
generated

Change-Id: I5ccd5911903fe5c2d577935eabbfc662ff87dec0