[banjo][usb] Begin transition to new banjo based USB protocol
We are doing a soft transition, so we retain support for the old USB protocol.

The USB bus driver now creates two devices for each USB device,
one that supports the new protocol and one that supports the old.
And we now have two copies of the usb-composite driver,
one for each version of the protocol.
Similarly we have two copies of the USB protocol support library.

This allows us to run with some USB drivers using the old protocol and
some using the new protocol simultaneously. After all the USB drivers
are switched to the new protocol we can remove support for the old one.

In this CL, all USB drivers are still using the old USB protocol.
However this has been tested with the usb-hid driver using the new protocol,
which will be submitted as a separate CL.

TEST: USB works on NUC.

Change-Id: Ibfcc475ea6d7bddc5d1f628d355c4a65e84ab7c9