[debugger] Support setting values.
Adds support for the "=" token.

Tokenization support for equality comparison is added (==) but will
issue an error. This is to prevent surprising error messages if somebody
uses that token today.

This is hooked up to initial support for binary operators. Currently
this is the only supported binary operator but the implementation is
generalized somewhat to support future additions.

The mock ExprNode implementation is separated out from the
expr_node_unittest to a separate file to it can be shared by the
operator unit test.

Adds support for all C++ operator precedence values in expr_parser.cc.
Most are unused and commented out to prevent warnings, but this makes
adding new ones less confusing.

TEST=unit,manual

Change-Id: Ib629a471cf658fd62123871362fd022c8f6d8495