[modular] fix: intent filters should be optional.
Because they were marked as non-optional (in ModuleManifest fidl
struct), but optional in the .cmx manifest, the FIDL calls to story
shell (which involve passing the ModuleManifest) have been failing
because of a non-nullable vector which was null.

This fix makes fuchsia::modular::ModuleManifest.intent_filters optional.

TEST=manually saw that tests were not reporting story shell FIDL errors.
A comprehensive test CL after this unbreaks people.

Change-Id: Id68eabefbc944f9bfe448b6c28fe88d178a19206