[debugger] Add first pass of casting support.
Adds a coersion function to convert values from one type to another.
This attempts to be as permissive as possible. It will be used in to
assign values to variables, and like GDB, we will basically trust the
programmer and do any type of cast that makes sense.

This is not yet hooked up to anything. In the future it will be used to
implement explicit casts in expressions, but that will reqiure more
enhancements like proper conversion to base classes when casting
pointers (which involves changing the value of the pointer).

TEST=unit

Change-Id: Ief4360d334d4a9c8660734c90bcc97a2fbfef92d