[mediaplayer] show black frame when flushing with hold_frame false
When a video renderer flushes, it either flushes all packets/images
or all packets/images other than the most recently presented one.
This is controlled by a 'hold_frame' flag on the flush call. The
last frame is held so the user sees the correct image during some
transitions. During other transitions, the last frame is not held
so the user sees black and all packets are freed.

The only way to get all frames released from an image pipe is to
present a different frame that needn't be released. This CL has the
video renderer creating a black frame for this purpose and presenting
it when a flush occurs with hold_frame == false.

QA-539 was occurring, because the video renderer was waiting for all
packets to be freed when hold_frame == false. This would not occur,
because the video renderer had no black frame with which to release
the last non-black frame.

The fake image pipe required an upgrade so it understands the use of
the black frame, which has different characteristics than the other
frames.

A new regression test was added for QA-539. This test fails without
the video renderer fix and succeeds with it.

TEST: added CQ regression test for QA-539

MTWN-203 #done
QA-539 #done

Change-Id: I73e4dba514eee5827ea9e7fac39d83d485b897a1