[mediaplayer] ensure VmoPayloadAllocator outlives its buffers
This CL add two RefPtr captures to ensure that the payload buffers
created by a VmoPayloadAllocator don't outlive the allocator itself
and call back into a deleted allocator. This was manifesting as a
deadlock as the callback attempted to take a lock. Repro rate was
low and only happened when the player to alternately play two files.

TEST: set_root_view mediaplayer_test_util --loop <file1> <file2>
Change-Id: Ia8fa6d8863acee88b55bd6384dcdbbee3a500fb4