[inspect] Fix bug in VMO Heap
We were not properly checking that merging is possible between a block
and its buddy. When two blocks of order n are merged into order n+1, we
check the buddy of the new block for continued merging. We, however,
need to make sure to check that the new buddy is of the right order
before merging.

This caused subtle errors when the ordering of Free operations was such
that a block gets merged but what would be its new buddy is split to
blocks of order < n+1.

TEST=runtests -t inspect-test

Change-Id: Ic4180b900635ede1701af1170cf12e30c2e80b1b