[debugger] Remove overly-zealous assertion.
I hit this assertion in a valid case. The assertion is removed and a
comment is added since getting an exception while we think the thread is
suspended is surprising.

This race condition isn't harmful since the kernel will handle resuming
from the exception and the suspend in either order.

TEST=manual

Change-Id: If0cf145b53c73738a065f916173ee5acd17c3879