[fit] Fix compatibility issue with GCC and C++ 17.
fit::is_null triggered GCC warnings for always-true comparison.
These warnings are harmless and impractical to remove so we suppress
them (as previously done for its libfbl equivalent).

In order to test this change, we had to modify the test suite so
that it can be compiled for userspace Zircon (not just for the host).
Some tests rely on libc++ symbols so they can't be compiled yet.
In particular, many of the examples use types such as std::string
for realism.  Move the examples out to separate files so we can
make progress compiling more of the actual tests.

Bug: ZX-3195
Test: fit-test
Change-Id: If8c3dfc4836ae3ac66d890684685e9bdfff075a8