[debugger] Add a write memory IPC call.
Plumbs the ability to write memory from the debug_agent to the symbol
data providers. This will be exercised in a future patch to support
assignment during expression evaluation.

Adds a missing "Job" implementation to the Minidump backend. Otherwise,
accidentally executing a job command with a dump open will crash.

TEST=unit

Change-Id: Ibc5c5d55185a1cae78697f47078d76445173f670