[zircon][crash] simplfiy crash handling
Since we are willing to give a restricted root
job to svchost, crash handling can be greatly
simplified. The JIRA bug below has a diagram,
but the TLDR is that crashsvc and devmgr were
doing very little other than redirecting.

- Remove crashsvc process and all that code
  except for the core that moves to svchost
- Remove code in devmgr that acts as middleman
  for crashes
- Remove the crashanalyzer FIDL svc code
  use the inspector library directly.

The logic of selecting which analyzer to use
is kept as before.

ZX-3199 #comment progress

TEST=manual: use $crasher and inspect that
     dumps are seen in the log output, otherwise
     unittests suffice

Change-Id: I84b24c85437ea0414e25e3fdabd53b55bb730257