[nandpart][test] Fix flaky nandpart-broker-test
There is a race between spawning a new ram-nand backed device and trying
to open it afterwards.

The correct fix would be to use fdio_watch_directory to wait for the
file to appear, but unfortunately due to a bug, openat always seems to
fail. The best we can do for now is just put a small sleep.

Tested: runtest -t nandpart-broker-test
Change-Id: Id281f48e935ece84934ac6628a40a9736df40933