[wlan][ath10k] Remove duplicate peer creation in client role.
In the original code, a peer would be added at both ath10k_mac_set_bss()
and ath10k_mac_bss_assoc(). The former one is necessary for management
packets, but the later one is duplicate. Also the type was wrong too.

Also remove the duplicate peer deletion code and add checking code to
detect the duplicate behaviors.

Also lower the first character of some output messages to be consistent
with other messages in the files. Adding some messages for future debug.

BUG=WLAN-85, WLAN-832, WLAN-597, WLAN-13
TESTED=run connect and disconnect commands to the following APs:
            fx-apple, fx-apple-2g, fx-att-pace, fx-aruba, Tou_eopn,
            fx-googlewifi, fx-netgear, fx-tplink, fx-ubi, GoogleGuest
       Also did a test to ensure the AP role is still working.

Change-Id: Ie6a8a6e64f9992814da6bdf408d2797f2b30f279