[debugger] Dim base classes in printed output.
Previously the debugger would highlight base classes with the "special"
annotation when printing a struct. This comes out as blue in the current
theme.

Using the debugger over the weekend showed that we should be trying to
hide these rather than calling attention to them. In normal usage, these
annotations aren't even very useful, and calling attention makes it
harder to skim for variable names.

This change makes the base class annotations dim to match the type
annotations.

TEST=manual

Change-Id: I552c443c3347b9e38e319ecdf7247dbb5953bbab