[stack unwind] Add libunwindstack.
This is a library from Android.

The contents of this directory are unchanged from upstream (does not include
the tests) and not hooked up to our build. Making it build in our context will
be done in a followup.

License and upstream location are described in README.fuchsia.

Pulled at revision ef6401091fbd047abcf08511f2d4d0b3a52c903f

OSRB-108
TEST=none

Change-Id: I61cc6ecc9f524a6c37177a299308a7938ecaacdf