[garnet] WlanFacade: add basic scan command
Add a basic scan command to WlanFacade to support sl4f.  This command
and support allows the caller to trigger a wlan scan and returns the
ssids found in the resulting scan.

Note: the corresponding WlanScanTest is part of a separate CL in ACTs.

Second note: also filed CONN-3 to track improved error handling in sl4f.

BUG: WLAN-529
Test: act.py -c basic_fuchsia.config -tc WlanScanTest
Change-Id: I2a3cbd3132fc79f01be107ca3b57b17b5456b3e3