[fidl] Visitor-style encoding.
As evidenced in
https://fuchsia-review.googlesource.com/c/zircon/+/227933,
utilizing the current FIDL buffer walker API to achieve tasks
outside of just encoding/decoding could require quite a bit of
wrangling. With llcpp, increasingly more use cases of FIDL traversal
are also popping up e.g. linearizing a message using coding tables.
This CL explores what could be the best interface design to
accommodate future consumers of the FIDL traversal logic.

As of this current design, the traversal logic is decoupled into
two pieces:

1. Walker, which travels along the FIDL coding table but does not
know about out-of-line offsets nor handle tables.
2. Visitor, which is called by the walker to patch pointers and handles,
but does not need to know about the coding table.

As a proof of concept, the FIDL encoder is re-implemented as a visitor,
while the decoder and the validator are kept unchanged. We could keep
these separate implementations and decide in the future if it is good
to switch over, or just implement handle closer, linearizers etc. as
visitors.

Test: CQ (in particular, fidl-test, fidl-simple-test)

Change-Id: Ib07ee472129236ce7ea0934dcadcbc5a3f43b100