[ledger] build merge bases with recursive merging
Ledger now uses recursive merging to construct a merge base, instead of
using the lowest unique common ancestor. The common ancestor algorithm
now return a set of lowest common ancestors that need to be merged to
get an appropriate merge base. The merge resolver works step-by-step:
it uses as many already existing merges as possible (using the index of
merges), then does one sub-merge. When one of the merged commits
contains a subset of the changes of the other, the conflict resolver is
not called.

TEST=almost all of common_ancestor_unittest, merge_resolver_unittest

Change-Id: I0e63a20a69e87b843b52e5e300c6ca3be108d05d