[ledger] index merge commits by their parents
In order to implement recursive merges, we need a way to retrieve merge
commits of two given commits. This is done by storing keys of the form
merges/<parent_with_lowest_id>/<parent_with_highest_id>/<merge_commit>
in the page database, and adding a method in PageStorage to retrieve
merge commits. This does not break compatibility: if some merge commits
are not in the database, they will simply be created again if needed,
and this time they will be added to the index.

TEST=added tests to page_storage_unittest, page_db_unittest

Change-Id: Ic042d1bd348b70ec24185f11c1db33b3a6226b8d