[sdk][doc] Recommend passing -x flag to amber_ctl add_srcs
By default, the amber_ctl add_srcs command disables all other updates
sources. This produces a confusing result if a developer wants to
publish their own packages but then finds that other packages on the
system no longer update.  The '-x' flag tells amber_ctl to not disable
other sources.

This flag can produce a different type of problem if the new source
provides a set of packages that overlap with other sources, but that's
unlikely in this scenario.

Test: docs
Change-Id: I326e92fd5bcc768afe3d0ac340c65c2e549ab9f0