[wlan][ath10k] Reconstruct QoS and 4-addr frames on QCA9880
The firmware delivers incoming frames in the "nwifi" format, with QoS
Control field stripped and addresses possibly rearranged. For mesh, we
need to reconstruct the original header. The driver code refers to this
process as "undecapping".

Note that this change should not affect 6174 since it goes through a
different RX code path.

TEST=Manually verified that 4-address mesh data frames are being
delivered correctly now. Smoke test on 9880 and 6174.

Change-Id: Ia1b435161a3b48a3995ea13f157faca2afadd40a