[kernel][locks] Add annontations for dump_thread* routines.
Add thread annotations for the following debug routines...
++ dump_thread
++ dump_all_threads
++ dump_thread_user_tid

Specifically, for each of these routines, make sure there are three
versions of the routine.  The base version (which demands that the
thread lock not be held, and will automatically obtain the lock), a
_locked version (which demands that the lock be held), and a
_during_panic version (which obtains no locks by going directly to the
_locked verion and skipping the compile time analysis)

Also, clean up the users in debug.cpp and x86/faults.cpp to use the
proper versions of each.

Test: Build and k thread (list|list_full) on QEMU
Change-Id: I88e3656e0ec64faf052a9efa32c1a20848456844