Addresses [shell scripts may fail if 'cd' writes to stdout](https://fuchsia.atlassian.net/browse/ZX-3212)
Updated all //scripts/* scripts using cd to find a path, to suppress
output, and calls to vars.sh now check for an error status.

Output is suppressed by redirecting stdout to /dev/null.
stderr is also redirected to
/dev/null to ensure there are no unintended side effects of calling
"cd" in a subshell (avoids unintentional escape sequences, such as
updating a terminal window title bar to the wrong directory).

Note that I found multiple ways the default (builtin) version of cd
can output to stdout. GNU bash can output the path to stdout in
some situations if CDPATH is set. zsh can output the path to stdout
even if CDPATH is not set.

Change-Id: I47da2bac4decd7c7ee8a325f9385d0fc4060aabe