[sdk][bazel] Remove safeguard for the main package's dilp file.
The kernel compiler used to not generate a dilp file for the main
package, instead generating a %{name}_kernel.dil-main.dilp file.
Declaring that file used to throw bazel errors as it wasn't generated.
This safeguard is no longer necessary as the dilp files generation is
wrapped under a script file that checks for existence and emptiness and
is safe to remove.

Bug: DX-851
Test: all tests pass, example dart_app run on target.
Change-Id: I48184a5ec1774c4e1125fdae7d63411fdcc63d09