[unwind] Port libunwindstack to Fuchsia.
This does a minimal port to Fuchsia of the libunwindstack library from
Android.

The bigger parts of the port are implemented in separate .cpp files in
the fuchsia/ directory. In a few places local modifications were made,
and these were all wrapped in __Fuchsia__ preprocessor defines. The
plan is to upstream these changes so we don't need to maintain any
modified files.

This is not currently hooked up to anything. We will experiment with it
in the debug_agent in future passes.

TEST=none

Change-Id: I3edffacc536fccefe472a123a2d4bcb91289b529